MicroServices/service_search_google/requirements.txt
googlesearch-python
flask
gunicorn
json-logging-py
